+++ CHANGELOG.md
@@ -4,8 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.1-beta.1]
+## [0.0.1-beta.2]
+### Changed
+- Clean up code in `Export-JsonManifest`
 
+### Added
+- mc client upload to S3
+- `Computer-UrlVTScan` function added for use with Drivers
+
+## [0.0.1-beta.1]
 ### Removed
 - Removed `-ApiBaseURI` from `Export-JsonManifest`